Factor out dateCmd dpll and dev info fetchers
