+++ pkg/fetcher/builder.go
@@ -5,7 +5,7 @@ package fetcher
 import (
 	"fmt"
 
-	"github.com/redhat-partner-solutions/vse-sync-testsuite/pkg/clients"
+	"github.com/redhat-partner-solutions/vse-sync-collection-tools/pkg/clients"
 )
 
 type AddCommandArgs struct {
